requirements.txt
flask
pytest
pytest-cov
pytest-randomly
pyhamcrest
